+++ apx_sdk/CHANGELOG.md
@@ -10,6 +10,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Change text representation of deliverer HTTP error to "HTTP error {code}".
 - Make `is_actor`, `is_activity`, `is_collection` and `is_object` compatible with FEP-2277.
+- Renamed `JsonSigner` type to `VerificationMethod`.
+- Renamed `signer` field on `JsonSignatureData` type to `verification_method`.
 
 ## [0.7.0] - 2024-12-19